chore: no circular deps
