+++ src/organizations/index.ts
@@ -1,24 +1,14 @@
 import { createSelector } from "@reduxjs/toolkit";
-import { call, select } from "saga-query";
 
-import { AuthApiCtx, authApi, cacheTimer } from "@app/api";
-import { exchangeToken } from "@app/auth";
 import { defaultEntity } from "@app/hal";
 import {
   createAssign,
   createReducerMap,
   createTable,
 } from "@app/slice-helpers";
-import { selectToken } from "@app/token";
-import {
-  ApiGen,
-  AppState,
-  HalEmbedded,
-  LinkResponse,
-  Organization,
-} from "@app/types";
+import { AppState, LinkResponse, Organization } from "@app/types";
 
-interface OrganizationResponse {
+export interface OrganizationResponse {
   address: string;
   city: string;
   created_at: string;
@@ -131,51 +121,6 @@ function deserializeOrganization(o: OrganizationResponse): Organization {
     name: o.name,
   };
 }
-
-type FetchOrgCtx = AuthApiCtx<
-  {},
-  HalEmbedded<{ organizations: OrganizationResponse[] }>
->;
-export const fetchOrganizations = authApi.get<FetchOrgCtx>(
-  "/organizations",
-  { saga: cacheTimer() },
-  function* onFetchOrgs(ctx, next) {
-    yield next();
-    if (!ctx.json.ok) {
-      return;
-    }
-
-    const orgs = ctx.json.data._embedded.organizations.sort(
-      (a, b) =>
-        new Date(b.updated_at).getTime() - new Date(a.updated_at).getTime(),
-    );
-    if (orgs.length > 0) {
-      ctx.actions.push(setOrganizationSelected(orgs[0].id));
-    }
-  },
-);
-
-interface CreateOrg {
-  name: string;
-}
-
-export const createOrganization = authApi.post<CreateOrg, OrganizationResponse>(
-  "/organizations",
-  function* onCreateOrg(ctx, next): ApiGen {
-    const { name } = ctx.payload;
-    ctx.request = ctx.req({
-      body: JSON.stringify({ name }),
-    });
-    yield next();
-    const token = yield* select(selectToken);
-    if (!ctx.json.ok) {
-      return;
-    }
-
-    yield* call(exchangeToken.run, exchangeToken(token));
-    ctx.actions.push(setOrganizationSelected(ctx.json.data.id));
-  },
-);
 
 export const entities = {
   organization: defaultEntity({
